Rename models.py into task_db.py for clarity
